requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
Pillow==10.2.0
numpy==1.24.3
rembg==2.0.54
gunicorn==21.2.0
onnxruntime==1.16.3 